player can die

+++ src/ooga/model/gameobjects/Player.java
@@ -82,7 +82,7 @@ public class Player extends Destroyable {
    * @param increment
    */
   @Override
-  public void incrementHealth(int increment) {
+  public void incrementHealth(Double increment) {
     int prevHealth = getHealth();
     int prevLives = getLives();
 
@@ -113,7 +113,7 @@ public class Player extends Destroyable {
    * @param x
    * @param y
    */
-  public void scaleVelocity(double x, double y) {
+  public void scaleVelocity(Double x, Double y) {
     Vector newVelocity = userMovement.getVelocity().multiply(new Vector(x, y));
     userMovement.setVelocity(newVelocity);
   }
@@ -127,11 +127,11 @@ public class Player extends Destroyable {
     return new ArrayList<>(activePowerUps);
   }
 
-  private void scaleSize(double scaleFactor) { size.scaleSize(scaleFactor); }
+  private void scaleSize(Double scaleFactor) { size.scaleSize(scaleFactor); }
 
-  private void incrementScore(double increment) { score += increment; }
+  private void incrementScore(Double increment) { score += increment; }
 
-  private void incrementLife(double increment) { lives += increment; }
+  private void incrementLife(Double increment) { lives += increment; }
 
   private void onKeyPress() { }
 }
